requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
requests
python-dotenv
schedule
loguru
fal-client
httpx
flask
flask-wtf
flask-limiter
flask-login
flask-session
Pillow
gunicorn